module46/requirements.txt
Django==4.1.13
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
rest_framework_simplejwt==0.0.2
pymongo==3.12.3
djongo==1.3.6
pytz==2024.1
pillow==10.3.0

